+++ __tests__/di.js
@@ -1,30 +1,18 @@
-import { DI, Component } from '../src/shamUI';
-import { renderApp } from './helpers';
-
-afterEach( () => {
-    DI.bind( 'service', undefined );
-} );
-
-function makeBinding( onRender ) {
-    return function() {
-        class Label extends Component {
-            render() {
-                if ( undefined !== onRender ) {
-                    onRender();
-                }
-                this.container.innerHTML = 'label text';
-            }
-        }
-        new Label( {
-            ID: 'simple-label-component-text',
-            container: document.querySelector( 'body' )
-        } );
-    };
-}
+import { DI, Component, start } from '../src/index';
 
 it( 'without component-binder', () => {
     const render = jest.fn();
-    renderApp( makeBinding( render ) );
+    class Label extends Component {
+        render() {
+            render();
+            this.container.innerHTML = 'label text';
+        }
+    }
+    new Label( {
+        ID: 'simple-label-component-text',
+        container: document.querySelector( 'body' )
+    } );
+    start();
     expect( document.querySelector( 'body' ).innerHTML ).toBe( 'label text' );
     expect( render ).toHaveBeenCalledTimes( 1 );
 } );