+++ __tests__/options-without-decorators.js
@@ -1,4 +1,4 @@
-import { Component, configureOptions, DI } from '../src/shamUI';
+import { Component, configureOptions, DI } from '../src/index';
 import { renderComponent, expectRenderedText } from './helpers';
 
 class Label extends Component {
@@ -199,7 +199,7 @@ it( 'getter & setter together', () => {
     component.options.text = 'test text';
     expect( component._text ).toBe( 'test text' );
     expect( component.options.text ).toBe( 'test text' );
-    DI.resolve( 'sham-ui' ).render.ONLY_IDS( 'dummy' );
+    DI.resolve( 'sham-ui:store' ).findById( 'dummy' ).render();
     expect( document.querySelector( 'body' ).textContent ).toBe( 'test text' );
 } );